refactor: correct renaming
